requirements.txt
streamlit
langchain
langchain-openai
langchain-mongodb
python-dotenv
Chroma
